move jobs into separate directory and return cleanup function
