+++ test-fixtures/configs/config.hcl
@@ -1,0 +1,9 @@
+# Copyright (c) HashiCorp, Inc.
+# SPDX-License-Identifier: MPL-2.0
+
+vault_addrs = "https://127.0.0.1:8200"
+vault_token = "sometoken"
+duration = "30s"
+report_mode = "terse"
+random_mounts = true
+nvalid = "option"
